requirements.txt
flask
flask_sqlalchemy
flask-classy
pyjade
requests
nose
python-coveralls
coverage